requirements.txt
pandas==2.2.3
numpy==2.1.2
plotly==5.24.1
flask==3.0.3
dash==3.0.1
dash-bootstrap-components==2.0.0
dash_core_components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
sentence_transformers
faiss-cpu
gunicorn
python-dotenv